ferramentas/legal-doc-assembler/pyproject.toml
[project]
name = "legal-doc-assembler"
version = "1.0.0"
description = "Deterministic engine for generating Brazilian legal documents from JSON data"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Claude Code Projetos" }
]
keywords = ["legal", "documents", "docx", "jinja2", "brazilian", "templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "docxtpl>=0.18.0",
    "jinja2>=3.1.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]
